+++ lib/tzispa_utils.rb
@@ -1,7 +1,7 @@
 module Tzispa
   module Utility
 
-
+    require 'tzispa/utils'
 
   end
 end