+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 Tzispa Utils
 
+## v0.1.4
+- Added CSVFixer utility class
+
+## v0.1.3
+- Fixes in String methods camelize and underscore
+
 ## v0.1.2
--- Added alias TzString for Tzispa::Utils::String
+- Added alias TzString for Tzispa::Utils::String
 
 ## v0.1.1
 - Added Tzispa::Utils::String